backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyjwt==2.8.0
httpx==0.27.0
pydantic==2.7.1
python-multipart==0.0.9
starlette==0.37.2
